requirements.txt
numpy==1.26.4; python_version < "3.14"
numpy==2.4.3; python_version >= "3.14"
opencv-python==4.10.0.84
PyYAML==6.0.3
Pillow==12.1.1
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
python-multipart==0.0.22
pytest==9.0.2
httpx==0.28.1
